+++ passes/R019/testdata/src/a/schema/schema.go
@@ -1,0 +1,7 @@
+package schema
+
+type ResourceData struct{}
+
+func (d *ResourceData) HasChanges(keys ...string) bool {
+	return false
+}